Rename 'host' to 'origin' and update server decorator
The 'host' field in the Prisma schema has been renamed to 'origin' for greater precision. In the server decorator, functionality has been added to handle the retrieval of the server id based on the origin in the request headers. Error handling is also present now, with an exception thrown when the server cannot be found. The unnecessary 'server' parameter in the 'resolve_settings' function in the server resolver file has been removed.