Refactor server modules and exception handling
This commit refactors various services and modules, improving code management. It adjusts the usage of PrismaService in param decorators, opting for server-based reliance. Exception handling within filters is also enhanced for better efficiency. Lastly, the bootstrapping of the Nest application is adjusted to utilize global modules and filters. This includes updates to GraphQL modules and switching local PrismaService usage to nestjs-prisma for enhanced code uniformity and efficiency.